src/main/java/kitchenpos/menus/tobe/domain/NewProduct.java
package kitchenpos.menus.tobe.domain;

import kitchenpos.common.domain.Price;

import javax.persistence.*;
import java.math.BigDecimal;
import java.util.Objects;
import java.util.UUID;

@Entity
@Table(name = "product")
public class NewProduct {
    @Column(name = "id", columnDefinition = "binary(16)")
    @Id
    private UUID id;

    @Embedded
    @Column(name = "price", nullable = false)
    private Price price;

    protected NewProduct() {
    }

    private NewProduct(UUID id, Price price) {
        this.id = id;
        this.price = price;
    }

    public static NewProduct create(UUID id, Price price) {
        return new NewProduct(id, price);
    }

    public UUID getId() {
        return id;
    }

    public Price getPrice() {
        return price;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        NewProduct that = (NewProduct) o;
        return Objects.equals(id, that.id);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }
}
